Add a new `autoExposedTransaction` property in `DatabaseClientConfig` and `getFieldExpressionSetWithExposedTransaction` parameters to some query functions, reverting the deprecation of the functions transforming rows using Vert.x's `PreparedQuery.mapping`
The deprecations are reverted because `FieldSet.getFieldExpressionSet` is the actual function that requires a transaction and `transaction` can be controlled in a finer-grained way by just wrapping this function with `transaction`. And therefore, wrapping Kotlin's `map`in `transaction` isn't the finest-grained way.

A slight performance issue that `executeQuery` calls `getFieldExpressionSet` when transforming every row is fixed.

Already reviewed locally by Copilot and improved.