Compare the code in the `oracle` and `mssql` modules with that in the `postgresql` module and fix some inconsistencies
This is done by first deleting the `oracle` and `mssql` modules, then copying and renaming the `postgresql` module to replace them, and them compare the changes line by line with the IntelliJ IDEA Commit Tool Window.

Confirmed that there is no Vert.x Unix domain socket support for Oracle and SQL Server, and that there are no `OraclePoolOptions` and `MSSQLPoolOptions` classes analogous to `PgPoolOptions` and `MySQLPoolOptions`.